+++ sketch-compose-core/src/main/kotlin/com/github/panpf/sketch/compose/AsyncImagePainter.kt
@@ -19,6 +19,7 @@ import androidx.compose.foundation.Image
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.NonRestartableComposable
 import androidx.compose.runtime.Stable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableFloatStateOf
@@ -71,6 +72,7 @@ import com.github.panpf.sketch.request.DisplayRequest
  *  destination.
  */
 @Composable
+@NonRestartableComposable
 fun rememberAsyncImagePainter(
     imageUri: String?,
     sketch: Sketch,
@@ -119,6 +121,7 @@ fun rememberAsyncImagePainter(
  *  destination.
  */
 @Composable
+@NonRestartableComposable
 fun rememberAsyncImagePainter(
     imageUri: String?,
     sketch: Sketch,
@@ -166,6 +169,7 @@ fun rememberAsyncImagePainter(
  *  destination.
  */
 @Composable
+@NonRestartableComposable
 fun rememberAsyncImagePainter(
     request: DisplayRequest,
     sketch: Sketch,